- adding DataBuffer, File and Movie to the C API - adding explicit constants for AP4_Track::Type enum - removing GetMetaData from AP4_Movie (was not implemented and redundant with AP4_File::GetMetaData) - returning AP4_Duration as opposed AP4_UI32 for AP4_Movie::GetDurationMs

+++ Source/Python/bento4/streams.py
@@ -16,17 +16,17 @@ class ByteStream(object):
     
     @property    
     def size(self):
-        v = c_ulong()
+        v = c_uint()
         res = lb4.AP4_ByteStream_GetSize(self.stream, byref(v))
         if res != 0:
             raise StreamException(res)
         return v.value
         
     def read_partial(self, bytes_to_read):
-        bytes_read = c_ulong()
+        bytes_read = c_uint()
         p = create_string_buffer(bytes_to_read)
         res = lb4.AP4_ByteStream_ReadPartial(self.stream, p,
-                                             c_ulong(bytes_to_read),
+                                             c_uint(bytes_to_read),
                                              byref(bytes_read))
         if res != 0:
             raise StreamException(res)
@@ -35,7 +35,7 @@ class ByteStream(object):
     def read(self, bytes_to_read):
         p = create_string_buffer(bytes_to_read)
         res = lb4.AP4_ByteStream_Read(self.stream, p,
-                                      c_ulong(bytes_to_read))
+                                      c_uint(bytes_to_read))
         if res != 0:
             raise StreamException(res)
         return p.raw
@@ -55,14 +55,14 @@ class ByteStream(object):
         return v.value
     
     def read_ui32(self):
-        v = c_ulong()
+        v = c_uint()
         res = lb4.AP4_ByteStream_ReadUI32(self.stream, byref(v))
         if res != 0:
             raise StreamException(res)
         return v.value
     
     def read_ui24(self):
-        v = c_ulong()
+        v = c_uint()
         res = lb4.AP4_ByteStream_ReadUI24(self.stream, byref(v))
         if res != 0:
             raise StreamException(res)
@@ -84,15 +84,15 @@ class ByteStream(object):
 
     def read_string(self, length):
         p = create_string_buffer(length+1)
-        res = lb4.AP4_ByteStream_ReadString(self.stream, p, c_ulong(length+1))
+        res = lb4.AP4_ByteStream_ReadString(self.stream, p, c_uint(length+1))
         if res != 0:
             raise StreamException(res)
         return p.value
     
     def write_partial(self, buffer):
-        bytes_written = c_ulong()
+        bytes_written = c_uint()
         res = lb4.AP4_ByteStream_WritePartial(self.stream, c_char_p(buffer),
-                                              c_ulong(len(buffer)),
+                                              c_uint(len(buffer)),
                                               byref(bytes_written))
         if res != 0:
             raise StreamException(res)
@@ -100,7 +100,7 @@ class ByteStream(object):
 
     def write(self, buffer):
         res = lb4.AP4_ByteStream_Write(self.stream, c_char_p(buffer),
-                                       c_ulong(len(buffer)))
+                                       c_uint(len(buffer)))
         if res != 0:
             raise StreamException(res)
         
@@ -115,12 +115,12 @@ class ByteStream(object):
             raise StreamException(res)
     
     def write_ui32(self, value):
-        res = lb4.AP4_ByteStream_WriteUI32(self.stream, c_ulong(value))
+        res = lb4.AP4_ByteStream_WriteUI32(self.stream, c_uint(value))
         if res != 0:
             raise StreamException(res)
                                               
     def write_ui24(self, value):
-        res = lb4.AP4_ByteStream_WriteUI24(self.stream, c_ulong(value))
+        res = lb4.AP4_ByteStream_WriteUI24(self.stream, c_uint(value))
         if res != 0:
             raise StreamException(res)
         
@@ -142,12 +142,12 @@ class ByteStream(object):
     def copy_to(self, receiver, size):
         res = lb4.AP4_ByteStream_CopyTo(self.stream,
                                         receiver.stream,
-                                        c_ulong(size))
+                                        c_uint(size))
         if res != 0:
             raise StreamException(res)
     
     def seek(self, position):
-        res = lb4.AP4_ByteStream_Seek(self.stream, c_ulong(position))
+        res = lb4.AP4_ByteStream_Seek(self.stream, c_uint(position))
         if res != 0:
             raise StreamException(res)
     
@@ -157,7 +157,7 @@ class ByteStream(object):
             raise StreamException(res)
     
     def tell(self):
-        pos = c_ulong()
+        pos = c_uint()
         res = lb4.AP4_ByteStream_Tell(self.stream, byref(pos))
         if res != 0:
             raise StreamException(res)
@@ -169,7 +169,7 @@ class MemoryByteStream(ByteStream):
     @staticmethod
     def FromBuffer(buffer):
         """Factory method"""
-        buffer_size = c_ulong(len(buffer))
+        buffer_size = c_uint(len(buffer))
         stream = lb4.AP4_MemoryByteStream_FromBuffer(c_char_p(buffer),
                                                      buffer_size)
         return MemoryByteStream(stream=stream)
@@ -177,7 +177,7 @@ class MemoryByteStream(ByteStream):
     
     def __init__(self, size=0, stream=None):
         if stream is None:
-            stream = lb4.AP4_ByteStream_Create(c_ulong(size))
+            stream = lb4.AP4_ByteStream_Create(c_uint(size))
         super(MemoryByteStream, self).__init__(stream)
